practice-compare-robyn/fast-api/requirements.txt
fastapi[standard]==0.115.6
SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
loguru==0.7.2
python-dotenv==1.0.1
pyjwt==2.10.1
passlib==1.7.4
bcrypt==4.0.1